Ticket 49305 - Improve atomic behaviours in 389-ds
Bug Description:  Previously our usage of atomics has ranged from
"okay" with PR_Atomic, to dangerous with atomic's only being
partially used. Notablp, PR_Atomic is much slower than gcc's
__atomic_* operations, especially given the barrier options
avaliable in gcc for thread sync.

Fix Description:  This replaces most calles to PR_Atomic
with the correct __atomic types. Importantly, this also
changes the slapi_counter types to __ATOMIC_RELAXED, which
removes thread ordering constraints, and allows just "atomics"
to function.

https://pagure.io/389-ds-base/issue/49305

Author: wibrown

Review by: mreynolds (Thanks!)

+++ ldap/servers/slapd/daemon.c
@@ -922,7 +922,7 @@ void slapd_daemon( daemon_ports_t *ports, ns_thrpool_t *tp )
 	PRIntn num_poll = 0;
 	PRIntervalTime pr_timeout = PR_MillisecondsToInterval(slapd_wakeup_timer);
 	PRThread *time_thread_p;
-	int threads;
+	uint64_t threads;
 	int in_referral_mode = config_check_referral_mode();
 	int connection_table_size = get_configured_connection_table_size();
 	the_connection_table= connection_table_new(connection_table_size);
@@ -1577,7 +1577,7 @@ handle_timeout( void )
 
 	if ( difftime(curtime, prevtime) >= 
 		slapd_housekeeping_timer ) {
-		int	num_active_threads;
+		uint64_t num_active_threads;
 
 		snmp_collator_update();
